+++ CHANGELOG.md
@@ -5,6 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## [Unreleased]
+
+### Fixed
+
+- Fix "Access Denied" on export generation caused by `User::getEntityID()` always returning -1
+
 ## [2.6.3] - 2026-08-04
 
 ### Fixed